Resolve "Material manage improve"

Closes #138

See merge request projectcms/cms-web!52

+++ src/actions/auth.ts
@@ -1,9 +1,10 @@
 import { ThunkAction, ThunkDispatch } from 'redux-thunk'
 import { AnyAction } from 'redux'
 import { setLoading, setErrorMessage } from '../actions/status'
-import { signIn } from '../apis/authorizations'
+import { signIn, signOut } from '../apis/authorizations'
 import { UPDATE_AUTH, DELETE_AUTH, SET_WECHAT_SHARE_URL } from '../constants/ActionTypes';
 import { SetWechatShareUrl, UpdateAuth, DeleteAuth } from '../types/auth';
+import { handleAxiosAsyncError } from '../utils/helper/errorHandle/axiosError';
 
 export const updateAuth = (accessToken: string): UpdateAuth => ({
   type: UPDATE_AUTH,
@@ -43,9 +44,13 @@ export const logout = (): ThunkAction<Promise<void>, {}, {}, AnyAction> => {
   return async (dispatch: ThunkDispatch<{}, {}, AnyAction>): Promise<void> => {
     try {
       dispatch(setLoading(true))
-      dispatch(deleteAuth())
+      const deleteAuthResult = await signOut()
+      if (deleteAuthResult) {
+        dispatch(deleteAuth())
+      }
       dispatch(setLoading(false))
     } catch (err) {
+      handleAxiosAsyncError(err)
       dispatch(setLoading(false))
     }
   }
